feat: split decorated backend code
JIRA: TNT-1700
